+++ SurfacePath.h
@@ -17,7 +17,7 @@
 
         ///
         void print() const {
-            for (int i=0; i<points.size(); i++)
+            for (size_t i=0; i<points.size(); i++)
                 printf("PointIdx: %d  isFix: %d\n", points[i], isFix[i]);
         }
